refactor(package-manager): move logic from separated files
{npm,pnpm}.service.ts logic move to package-manager.service.ts
simplify code
improve tests for future development
