+++ packages/cli/src/services/initializer.service.ts
@@ -1,7 +1,7 @@
 import { DependencyTypeEnum } from '../type/enums/dependency-type.enum.js';
 import type { PackageInterface } from '../type/interfaces/package.interface.js';
 import type { PackagesDependencyGroup } from '../type/types/packages-dependency-group.interface.js';
-import { PackageManagerService } from './package-managers/package-manager.service.js';
+import { PackageManagerService } from './package-manager/package-manager.service.js';
 
 export class InitializerService {
   packages: PackageInterface[] = [];
