+++ packages/cli/src/type/enums/installation-type.enum.ts
@@ -1,0 +1,5 @@
+export enum InstallationType {
+  dependency = '',
+  developmentDependency = '-D',
+  global = '--global',
+}
