--- packages/cli/src/type/interfaces/package-manager.interface.ts
@@ -1,14 +1,0 @@
-import { DependencyTypeEnum } from '../enums/dependency-type.enum.js';
-import type { PackagesEnumKeys } from '../enums/packages.enum.js';
-import type { PackagesDependencyGroup } from '../types/packages-dependency-group.interface.js';
-
-export interface PackageManager {
-  readonly installationType: {
-    [DependencyTypeEnum.dependency]: string;
-    [DependencyTypeEnum.devDependency]: string;
-    [DependencyTypeEnum.global]: string;
-  };
-  install(packages: PackagesDependencyGroup): Promise<void>;
-  uninstall(packages: string[]): Promise<void>;
-  create(packages: PackagesEnumKeys[]): Promise<void>;
-}
